Move useActiveModalFrontstage
